python/weather-demand-bot/requirements.txt
python-telegram-bot==21.3
pandas==2.2.3
requests==2.32.3
openpyxl==3.1.2
